Refactor ModelBuilder, core types into separate files
