+++ newton/core/__init__.py
@@ -13,17 +13,78 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .builder import ModelBuilder
 from .contact import Contact
 from .control import Control
-from .model import SHAPE_FLAG_COLLIDE_GROUND, SHAPE_FLAG_COLLIDE_SHAPES, SHAPE_FLAG_VISIBLE, Model
+from .model import Model
 from .state import State
+from .types import (
+    GEO_BOX,
+    GEO_CAPSULE,
+    GEO_CONE,
+    GEO_CYLINDER,
+    GEO_MESH,
+    GEO_NONE,
+    GEO_PLANE,
+    GEO_SDF,
+    GEO_SPHERE,
+    JOINT_BALL,
+    JOINT_COMPOUND,
+    JOINT_D6,
+    JOINT_DISTANCE,
+    JOINT_FIXED,
+    JOINT_FREE,
+    JOINT_MODE_FORCE,
+    JOINT_MODE_TARGET_POSITION,
+    JOINT_MODE_TARGET_VELOCITY,
+    JOINT_PRISMATIC,
+    JOINT_REVOLUTE,
+    JOINT_UNIVERSAL,
+    PARTICLE_FLAG_ACTIVE,
+    SDF,
+    SHAPE_FLAG_COLLIDE_GROUND,
+    SHAPE_FLAG_COLLIDE_SHAPES,
+    SHAPE_FLAG_VISIBLE,
+    JointAxis,
+    Mesh,
+    ModelShapeGeometry,
+    ModelShapeMaterials,
+)
 
 __all__ = [
+    "GEO_BOX",
+    "GEO_CAPSULE",
+    "GEO_CONE",
+    "GEO_CYLINDER",
+    "GEO_MESH",
+    "GEO_NONE",
+    "GEO_PLANE",
+    "GEO_SDF",
+    "GEO_SPHERE",
+    "JOINT_BALL",
+    "JOINT_COMPOUND",
+    "JOINT_D6",
+    "JOINT_DISTANCE",
+    "JOINT_FIXED",
+    "JOINT_FREE",
+    "JOINT_MODE_FORCE",
+    "JOINT_MODE_TARGET_POSITION",
+    "JOINT_MODE_TARGET_VELOCITY",
+    "JOINT_PRISMATIC",
+    "JOINT_REVOLUTE",
+    "JOINT_UNIVERSAL",
+    "PARTICLE_FLAG_ACTIVE",
+    "SDF",
     "SHAPE_FLAG_COLLIDE_GROUND",
     "SHAPE_FLAG_COLLIDE_SHAPES",
     "SHAPE_FLAG_VISIBLE",
     "Contact",
     "Control",
+    "JointAxis",
+    "Mesh",
     "Model",
+    "ModelBuilder",
+    "ModelShapeGeometry",
+    "ModelShapeMaterials",
     "State",
 ]
